LayoutTests/platform/ios-sim/svg/custom/getSubStringLength-expected.txt

PASS svgText.getSubStringLength(-1, 2) threw exception Error: IndexSizeError: DOM Exception 1.
PASS svgText.getSubStringLength(-1, 0) threw exception Error: IndexSizeError: DOM Exception 1.
PASS svgText.getSubStringLength(1, 3) is 40
PASS svgText.getSubStringLength(0, 4) is 60
PASS svgText.getSubStringLength(3, 0) threw exception Error: IndexSizeError: DOM Exception 1.
PASS svgText.getSubStringLength(0, 0) is 0
PASS svgText.getSubStringLength(2, 0) is 0
PASS svgText.getSubStringLength(0, 1) is 20
PASS svgText.getSubStringLength(1, 1) is 20
PASS svgText.getSubStringLength(2, 1) is 20
PASS svgText.getSubStringLength(0, 3) is 60
PASS svgText.getSubStringLength(1, -1) threw exception Error: IndexSizeError: DOM Exception 1.
PASS svgText.getSubStringLength(2, -1) threw exception Error: IndexSizeError: DOM Exception 1.
PASS svgText.getSubStringLength(3, -1) threw exception Error: IndexSizeError: DOM Exception 1.
PASS svgText.getSubStringLength(3, -3) threw exception Error: IndexSizeError: DOM Exception 1.
PASS successfullyParsed is true

TEST COMPLETE
